Remove global css
