--- assets/css/presenter.css
@@ -1,46 +1,0 @@
-.presenter {
-    width: 100vw;
-    height: 100vh;
-    overflow: hidden;
-    background-color: #eeeeee;
-    background-image: url("../img/web/Robot.png");
-    background-size: cover;
-    background-repeat: no-repeat;
-}
-
-.presenter .gameView {
-    position: absolute;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 0;
-    background-color: #aaaaaa;
-}
-
-.presenter .clientList {
-    position: relative;
-    top: 0;
-    z-index: 10;
-    background-color: rgba(255, 255, 255, 0.8);
-    width: 150px;
-    overflow: hidden;
-    height: auto;
-    padding: 8px;
-}
-
-.clientList li {
-    list-style-image: none;
-}
-
-#actionO_right > .actionO_Line {
-    background: none !important;
-}
-
-#actionO_left > .actionO_Line {
-    background: none !important;
-}
-
-.actionO_Name {
-    color: #ffa241 !important;
-    text-shadow: 4px 4px 8px #930804;
-}
