+++ index.html
@@ -5,7 +5,6 @@
     <meta charset="UTF-8">
     <title>RobotIDE</title>
     <link href="assets/css/main.css" rel="stylesheet">
-    <link href="assets/css/presenter.css" rel="stylesheet">
   </head>
   <body>
     <script src="node_modules/blockly/blockly_compressed.js"></script>
